Temp and i2cscan for 8266
Added the BME280 Temp/humidity/baro code for 8266.
Including a subscription to a correction channel in which a correction value for baro pressure can be posted.
This value is added to the measured value

I2Cscan is a simple sketch scanning the i2c addresses and reporting where something was found